Fehlerbehebung und Wartekreis bei Registrieren
