+++ App2Night/App2Night.csproj
@@ -130,9 +130,6 @@
     <Compile Include="Views\FensterAnmelden.xaml.cs">
       <DependentUpon>FensterAnmelden.xaml</DependentUpon>
     </Compile>
-    <Compile Include="FensterAnmOdReg.xaml.cs">
-      <DependentUpon>FensterAnmOdReg.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Views\FensterAnmOdReg.xaml.cs">
       <DependentUpon>FensterAnmOdReg.xaml</DependentUpon>
     </Compile>
@@ -193,10 +190,6 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Views\FensterAnmelden.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="FensterAnmOdReg.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
