--- App2Night/FensterAnmOdReg.xaml.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
-
-// Die Elementvorlage "Leere Seite" ist unter http://go.microsoft.com/fwlink/?LinkId=234238 dokumentiert.
-
-namespace App2Night.Views
-{
-    /// <summary>
-    /// Eine leere Seite, die eigenständig verwendet oder zu der innerhalb eines Rahmens navigiert werden kann.
-    /// </summary>
-    public sealed partial class FensterAnmOdReg : Page
-    {
-        public FensterAnmOdReg()
-        {
-            this.InitializeComponent();
-        }
-        private void Anmelden_wechselZuAnmelden(object sender, RoutedEventArgs e)
-        {
-            this.Frame.Navigate(typeof(FensterAnmelden));
-        }
-
-        private void Registrieren_wechselZuReg(object sender, RoutedEventArgs e)
-        {
-            this.Frame.Navigate(typeof(FensterReg));
-        }
-    }
-}
